+++ packages/sdk-redux/src/reduxSlices/transactionTrackerSlice/trackedTransaction.ts
@@ -27,4 +27,8 @@ export interface TrackedTransaction {
      * `true` when Subgraph polling was successful, i.e. Subgraph has indexed the transaction. Will be `undefined` if the polling was unsuccessful or we don't know whether it's in sync or not.
      */
     isSubgraphInSync?: true;
+    /**
+     * The block number where transaction succeeded. Useful for creating polling mechanisms outside of the SDK. For example, another Subgraph API slize.
+     */
+    blockTransactionSucceededIn?: number;
 }